requirements.txt
streamlit
openai
python-dotenv
requests
lightrag-hku
Pillow

